% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_zones.R
\name{plot_zones}
\alias{plot_zones}
\title{Plot illustrative blast zones}
\usage{
plot_zones(radii, labels = NULL)
}
\arguments{
\item{radii}{numeric vector (km)}

\item{labels}{optional vector of zone labels}
}
\description{
Plot illustrative blast zones
}
